feat(markdown): project markdown outputs in host
