+++ apps/notebook/src/renderer-plugins/isolated-renderer.js
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:9af022d01d6cb8dca7e33ec8f8516b53796da7af36918b48d4734b3aa18fdbd0
-size 1501177
+oid sha256:dbda7ea8adadd34d8ff646a76c2b4de1fbeb3d1803f230911bf228a0c220d217
+size 1530354
